Extract Container class to new module
